# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Perfume Inventory Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices:
    get:
      operationId: getInvoices
      tags: [invoices]
      summary: Get all invoices
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create invoice and deduct inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoice"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete invoice and restore inventory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /storage/uploads/request-url:
    post:
      tags: [storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /storage/objects/{objectPath}:
    get:
      tags: [storage]
      operationId: getStorageObject
      summary: Serve an uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
  /inventory/{id}/images:
    get:
      tags: [product-images]
      operationId: getProductImages
      summary: Get all images for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of product images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductImage"
    post:
      tags: [product-images]
      operationId: addProductImage
      summary: Add an image to a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductImage"
      responses:
        "201":
          description: Image added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImage"
  /inventory/{id}/images/{imageId}:
    delete:
      tags: [product-images]
      operationId: deleteProductImage
      summary: Delete a product image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Image deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    InvoiceItem:
      type: object
      properties:
        id:
          type: integer
        invoice_id:
          type: integer
        inventory_id:
          type: integer
          nullable: true
        barcode:
          type: string
          nullable: true
        brand:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
        size:
          type: string
          nullable: true
        concentration:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        qty:
          type: integer
        unit_price_aed:
          type: number
        cost_usd:
          type: number
          nullable: true
        created_at:
          type: string
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoice_number:
          type: string
        customer_id:
          type: integer
          nullable: true
        customer_name:
          type: string
          nullable: true
        date:
          type: string
        status:
          type: string
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        notes:
          type: string
          nullable: true
        created_at:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
    CreateInvoiceItem:
      type: object
      required:
        - inventory_id
        - qty
        - unit_price_aed
      properties:
        inventory_id:
          type: integer
        qty:
          type: integer
        unit_price_aed:
          type: number
    CreateInvoice:
      type: object
      required:
        - items
      properties:
        customer_id:
          type: integer
          nullable: true
        customer_name:
          type: string
          nullable: true
        date:
          type: string
          nullable: true
        discount:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateInvoiceItem"
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
    ProductImage:
      type: object
      properties:
        id:
          type: integer
        inventory_id:
          type: integer
        object_path:
          type: string
        caption:
          type: string
          nullable: true
        sort_order:
          type: integer
        created_at:
          type: string
    CreateProductImage:
      type: object
      required: [object_path]
      properties:
        object_path:
          type: string
        caption:
          type: string
          nullable: true
        sort_order:
          type: integer
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean